newadmin/requirements.txt
Django==1.8.2
django-appconf==1.0.1
django-bootstrap-form==3.2
django-compressor==1.5
django-debug-toolbar==1.3.0
django-extensions==1.5.5
django-localflavor==1.1
python-magic==0.4.6
requests==2.7.0
six==1.9.0
xlwt==1.0.0
pydicom==0.9.9